Move UserRepository to BlackjackService
