+++ src/main/java/kr/ac/cnu/web/service/BlackjackService.java
@@ -1,11 +1,15 @@
 package kr.ac.cnu.web.service;
 
+import kr.ac.cnu.web.exceptions.NoLoginException;
 import kr.ac.cnu.web.games.blackjack.Deck;
 import kr.ac.cnu.web.games.blackjack.GameRoom;
 import kr.ac.cnu.web.model.User;
+import kr.ac.cnu.web.repository.UserRepository;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.HashMap;
+import java.util.List;
 import java.util.Map;
 
 /**
@@ -17,7 +21,11 @@ public class BlackjackService {
     private final int MIN_BET = 3000;
     private final Map<String, GameRoom> gameRoomMap = new HashMap<>();
 
-    public GameRoom createGameRoom(User user) {
+    @Autowired
+    private UserRepository userRepository;
+
+    public GameRoom createGameRoom(String name) {
+        User user = this.getUserFromSession(name);
         Deck deck = new Deck(DECK_NUMBER);
 
         GameRoom gameRoom = new GameRoom(deck);
@@ -28,12 +36,13 @@ public class BlackjackService {
         return gameRoom;
     }
 
-    public GameRoom joinGameRoom(String roomId, User user) {
+    public GameRoom joinGameRoom(String roomId, String name) {
         // multi player Game 이 아니므로 필요가 없다.
         return null;
     }
 
-    public void leaveGameRoom(String roomId, User user) {
+    public void leaveGameRoom(String roomId, String name) {
+        User user = this.getUserFromSession(name);
         gameRoomMap.get(roomId).removePlayer(user.getName());
     }
 
@@ -41,8 +50,9 @@ public class BlackjackService {
         return gameRoomMap.get(roomId);
     }
 
-    public GameRoom bet(String roomId, User user, long bet) {
+    public GameRoom bet(String roomId, String name, long bet) {
         GameRoom gameRoom = gameRoomMap.get(roomId);
+        User user = this.getUserFromSession(name);
 
         if (bet < MIN_BET) {
             return gameRoom;
@@ -55,16 +65,18 @@ public class BlackjackService {
         return gameRoom;
     }
 
-    public GameRoom hit(String roomId, User user) {
+    public GameRoom hit(String roomId, String name) {
         GameRoom gameRoom = gameRoomMap.get(roomId);
+        User user = this.getUserFromSession(name);
 
         gameRoom.hit(user.getName());
 
         return gameRoom;
     }
 
-    public GameRoom stand(String roomId, User user) {
+    public GameRoom stand(String roomId, String name) {
         GameRoom gameRoom = gameRoomMap.get(roomId);
+        User user = this.getUserFromSession(name);
 
         gameRoom.stand(user.getName());
         gameRoom.playDealer();
@@ -72,12 +84,32 @@ public class BlackjackService {
         return gameRoom;
     }
 
-    public GameRoom doubleDown(String roomId, User user) {
+    public GameRoom doubleDown(String roomId, String name) {
         GameRoom gameRoom = gameRoomMap.get(roomId);
+        User user = this.getUserFromSession(name);
 
         gameRoom.doubleDown(user.getName());
         gameRoom.playDealer();
 
         return gameRoom;
+    }
+
+    public User login(String name) {
+        return this.getUserFromSession(name);
+    }
+
+    private User getUserFromSession(String name) {
+        return userRepository.findById(name).orElseThrow(() -> new NoLoginException());
+    }
+
+    public List<User> getTopNUsers(int nRecords) {
+        List<User> users = userRepository.findAllByOrderByAccountDesc();
+
+        if ((nRecords > 0) && (nRecords < users.size())) {
+            return users.subList(0, nRecords);
+        }
+        else {
+            return users;
+        }
     }
 }
